move grpc server logic into util
